refactor

+++ jdbc-store/src/main/java/com/kii/beehive/portal/jdbc/dao/TagThingRelationDao.java
@@ -5,6 +5,8 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.logging.log4j.util.Strings;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
 import com.kii.beehive.portal.jdbc.entity.TagIndex;
@@ -14,32 +16,36 @@ import com.kii.beehive.portal.jdbc.entity.TagType;
 @Repository
 public class TagThingRelationDao extends BaseDao<TagThingRelation> {
 
-	//private Logger log= LoggerFactory.getLogger(TagThingRelationDao.class);
+	private Logger log= LoggerFactory.getLogger(TagThingRelationDao.class);
 	
 	public static final String TABLE_NAME = "rel_thing_tag";
 	public static final String KEY = "id";
 	
 	public void delete( Long tagID, Long thingID){
-		String sql = "DELETE FROM " + this.getTableName() + " WHERE ";
-		
-		StringBuilder where = new StringBuilder();
-		List<Object> params = new ArrayList<Object>();
-		if(thingID != null){
-			where.append(TagThingRelation.THING_ID + " = ? "); 
-			params.add(thingID);
-		}
-		
-		if(tagID != null){
-			if(where.length() > 0){
-				where.append(" AND ");
+		if(tagID != null || thingID != null){
+			String sql = "DELETE FROM " + this.getTableName() + " WHERE ";
+			
+			StringBuilder where = new StringBuilder();
+			List<Object> params = new ArrayList<Object>();
+			if(thingID != null){
+				where.append(TagThingRelation.THING_ID + " = ? "); 
+				params.add(thingID);
 			}
-			where.append(TagThingRelation.TAG_ID+" = ? ");
-			params.add(tagID);
+			
+			if(tagID != null){
+				if(where.length() > 0){
+					where.append(" AND ");
+				}
+				where.append(TagThingRelation.TAG_ID+" = ? ");
+				params.add(tagID);
+			}
+			Object[] paramArr = new Object[params.size()];
+			paramArr = params.toArray(paramArr);
+			
+	        jdbcTemplate.update(sql+where.toString(),paramArr);
+		}else{
+			log.warn("tagID and thingID are null");
 		}
-		Object[] paramArr = new Object[params.size()];
-		paramArr = params.toArray(paramArr);
-		
-        jdbcTemplate.update(sql+where.toString(),paramArr);
 	}
 
 	@Override
@@ -79,16 +85,19 @@ public class TagThingRelationDao extends BaseDao<TagThingRelation> {
 	}
 	
 	public TagThingRelation findByThingIDAndTagID(Long thingID, Long tagID) {  
-		String sql = "SELECT * FROM " + this.getTableName() + " WHERE "+ TagThingRelation.THING_ID +"=? AND "+ TagThingRelation.TAG_ID + "=?";
-        List<Map<String, Object>> rows = jdbcTemplate.queryForList(sql, thingID,tagID);
-        List<TagThingRelation> list = mapToList(rows);
-        if(list.size() > 0){
-        	return list.get(0);
-        }
+		if(tagID != null && thingID != null){
+			String sql = "SELECT * FROM " + this.getTableName() + " WHERE "+ TagThingRelation.THING_ID +"=? AND "+ TagThingRelation.TAG_ID + "=?";
+	        List<Map<String, Object>> rows = jdbcTemplate.queryForList(sql, thingID,tagID);
+	        List<TagThingRelation> list = mapToList(rows);
+	        if(list.size() > 0){
+	        	return list.get(0);
+	        }
+		}
         return null;
     }
 
 	/**
+	 * @deprecated
 	 * get the tag thing relation by global thing id, tag type and tag display name <br>
 	 *     if tag display name is not specified, will get all the relations under the global thing id and tag type
 	 * @param globalThingID
